PAY-65: Add created at interface and trait + Adjust models to use the new trait
